improvements made to UI
