+++ app/src/main/java/orbital/raspberry/neighber/MainActivity2.java
@@ -36,13 +36,12 @@ public class MainActivity2 extends AppCompatActivity {
     private TextView browsereq, browseoff;
     private List<Post> posts;
     private ListView listViewRequests;
-    private EditText searchtxt;
-    private Button searchbtn;
+
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_main2);
+        setContentView(R.layout.activity_main);
 
         //Get Firebase auth instance
         auth = FirebaseAuth.getInstance();
@@ -63,14 +62,13 @@ public class MainActivity2 extends AppCompatActivity {
         browsereq.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                startActivity(new Intent(MainActivity2.this, MainActivity.class));
             }
         });
 
         browseoff.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-
+                startActivity(new Intent(MainActivity2.this, MainActivity.class));
             }
         });
 
@@ -111,10 +109,6 @@ public class MainActivity2 extends AppCompatActivity {
 
         //////////////////////End Navigation////////////////////////////
 
-        searchtxt = (EditText) findViewById(R.id.searchtxt);
-        searchbtn = (Button) findViewById(R.id.searchbtn);
-
-
         final DatabaseReference mDatabase = FirebaseDatabase.getInstance().getReference("posts");
 
         //attaching value event listener
@@ -131,15 +125,15 @@ public class MainActivity2 extends AppCompatActivity {
                     Post post = postSnapshot.getValue(Post.class);
 
                     //If post type is request aka 1
-                        if(post.getPosttype() == 2 && post.getStatus() == 1) {
+                    if(post.getPosttype() == 1 && post.getStatus() == 1) {
 
-                            String datetime = getDate(post.getTimestamp());
+                        String datetime = getDate(post.getTimestamp());
 
-                            post.setDatetime(datetime);
+                        post.setDatetime(datetime);
 
-                            //adding to the list
-                            posts.add(post);
-                        }
+                        //adding to the list
+                        posts.add(post);
+                    }
                 }
 
                 //creating adapter
@@ -158,7 +152,7 @@ public class MainActivity2 extends AppCompatActivity {
             @Override
             public void onItemClick(AdapterView<?> parent, View view, int position, long id) {
                 Post post = posts.get(position);
-                //Toast.makeText(MainActivity.this, "Item: " + request.getItemname() + " selected.", Toast.LENGTH_SHORT).show();
+                //Toast.makeText(MainActivity.this, "Item: " + post.getItemname() + " selected.", Toast.LENGTH_SHORT).show();
                 String requesterid = post.getUserid();
                 String postid = post.getPostid();
 
@@ -166,48 +160,6 @@ public class MainActivity2 extends AppCompatActivity {
                 i.putExtra("ruserid", requesterid);
                 i.putExtra("rpostid", postid);
                 startActivity(i);
-
-            }
-        });
-
-
-        searchbtn.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-
-                String query = searchtxt.getText().toString().trim();
-                int found = 0;
-
-                for(Post p : posts){
-
-                    String splitname[] = p.getItemname().split(" ");
-
-                    for(int i=0; i<splitname.length; i++) {
-
-                        if (splitname[i].equals(query)) {
-
-                            String requesterid = p.getUserid();
-                            String postid = p.getPostid();
-
-                            found = 1;
-
-                            Intent i2 = new Intent(MainActivity2.this, PostActivity.class);
-                            i2.putExtra("ruserid", requesterid);
-                            i2.putExtra("rpostid", postid);
-                            startActivity(i2);
-
-                            break;
-                        }
-                    }
-
-                    if(found == 1){
-                        break;
-                    }
-
-                }
-                if(found == 0) {
-                    Toast.makeText(MainActivity2.this, "No such items found", Toast.LENGTH_SHORT).show();
-                }
 
             }
         });
