+++ app/src/main/java/orbital/raspberry/neighber/WriteOfferActivity2.java
@@ -27,6 +27,7 @@ public class WriteOfferActivity2 extends AppCompatActivity {
     private int rrecordcount;
     private Button submitBtn;
     private EditText offerdescTxt;
+    private TextView itemnameTxt;
 
     private FirebaseAuth auth;
 
@@ -87,6 +88,7 @@ public class WriteOfferActivity2 extends AppCompatActivity {
 
         submitBtn = (Button)findViewById(R.id.submitRequest);
         offerdescTxt = (EditText)findViewById(R.id.offerdesc);
+        itemnameTxt = (TextView) findViewById(R.id.itemnametxt) ;
 
         //Get Firebase auth instance
         auth = FirebaseAuth.getInstance();
@@ -118,6 +120,8 @@ public class WriteOfferActivity2 extends AppCompatActivity {
                 Post post = dataSnapshot.getValue(Post.class);
 
                 ritemname = post.getItemname();
+
+                itemnameTxt.setText("Item: " + ritemname);
 
                 rrecordcount = post.getRecordcount();
