+++ app/src/main/res/layout/activity_write_offer.xml
@@ -74,17 +74,26 @@
 
     </LinearLayout>
 
+    <TextView
+        android:id="@+id/itemnametxt"
+        android:layout_width="fill_parent"
+        android:layout_height="wrap_content"
+        android:textSize="24dp"
+        android:hint="Item Name"
+        android:textStyle="bold"
+        android:layout_marginTop="32dp"
+        android:layout_alignParentTop="true"
+        android:layout_alignParentStart="true" />
 
     <TextView
         android:id="@+id/infotxt"
         android:layout_width="fill_parent"
         android:layout_height="wrap_content"
         android:textSize="24dp"
-        android:inputType="textPersonName"
         android:hint="Offer and Condition of Loan"
         android:textStyle="bold"
-        android:layout_marginTop="23dp"
-        android:layout_below="@+id/posttype"
+        android:layout_marginTop="37dp"
+        android:layout_below="@+id/itemnametxt"
         android:layout_alignParentStart="true" />
 
     <EditText
